+++ src/test_coef_skin.f90
@@ -42,7 +42,7 @@ PROGRAM TEST_COEF
 
    REAL(wp), DIMENSION(lx,ly) :: Cd, Ce, Ch, Cp_ma, rgamma
 
-   REAL :: zt, nu_air
+   REAL(wp) :: zt, nu_air
 
    CHARACTER(len=3) :: czt, czu
